requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
facebook-business==22.0.0
google-ads==26.0.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==4.5.1
celery==5.3.6
aioredis==2.0.1
prometheus-client==0.16.0
# Thêm các package cần thiết cho utilities
python-json-logger==2.0.7
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2